pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pllm_mve"
version = "0.1.0"
description = "Minimal Viable Experiment for Generative Active Preference Learning"
authors = [
    {name = "Research Team"},
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.57.0",
    "together>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "black>=23.3.0",
    "ruff>=0.0.270",
    "mypy>=1.3.0",
]

[tool.setuptools]
packages = ["pllm_mve"]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]
ignore = ["E501"]